Cancellable moves
This now use a CancelHook together with non-blocking
moves, to allow moves to be aborted. Potentially, we could
eliminate the stop action.

I've set the polling frequency to 50ms, which seems fine. For
timing-critical applications (or just places where cancellation
might be undesirable) there's a `block_cancellation` argument.

Blocking cancellation not only avoids raising an exception, it
also uses `notify_on_stop` to revert to the old (blocking)
behaviour on a per-move basis. I am confident this will have
no measurable effect on the timing of moves vs the old
blocking moves.

labthings-sangaboard is now incompatible with firmware < v1,
but this is unlikely to be a problem.