Propagate exceptions when moves are cancelled
If a move is cancelled, we should propagate the exception, so
that any actions that are calling the move method will also
be cancelled. For example, if an autofocus action is cancelled
during a move, not propagating the exception will allow it
to continue making the next move after this one aborts,
rather than cancelling the whole autofocus operation.

If different behaviour is desired, it's best to achieve that by
adding exception handling to the calling Action.